+++ google2/CHANGELOG.md
@@ -6,11 +6,12 @@ This file documents changes to the `workbench-google2` library, including notes 
 
 Added
 - Add `setBucketPolicyOnly`
+- Add `setObjectMetadata`
 
 Changed
 - Update Google Cloud Storage client library to 1.77.0
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.4-88966fd"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.4-TRAVIS-REPLACE-ME"`
 
 ## 0.3